refactor: refactor the dashboard component
