+++ frontend/src/views/DashboardView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import Dashboard from '@/components/Dashboard.vue';
+import Dashboard from '@/components/dashboard/Dashboard.vue';
 </script>
 
 <template>
